Added comma inside comment of Register.v line 1.
Thanks for noting the missing comma.  I noticed this was missing in the starter code and added the comma for testing and removed it afterwards, but I think including inside the comment tag and moving the comment tag up makes it easier for testing.  That was part of why I thought I might include the .vcd file to save the trouble of others missing the comma, etc.  This is better.